+++ tests/core/test_playback.py
@@ -708,6 +708,7 @@ class CorePlaybackWithOldBackendTest(unittest.TestCase):
         c = core.Core(mixer=None, backends=[b])
         c.tracklist.add(uris=['dummy1:a'])
         c.playback.play()  # No TypeError == test passed.
+        b.playback.play.assert_called_once_with()
 
 
 class TestPlay(unittest.TestCase):
@@ -727,3 +728,26 @@ class TestPlay(unittest.TestCase):
         self.core.playback.play(tlid=self.tl_tracks[1].tlid)
         self.backend.playback.change_track.assert_called_once_with(
             self.tl_tracks[1].track)
+
+
+class Bug1177RegressionTest(unittest.TestCase):
+    def test(self):
+        b = mock.Mock()
+        b.uri_schemes.get.return_value = ['dummy']
+        b.playback = mock.Mock(spec=backend.PlaybackProvider)
+        b.playback.change_track.return_value.get.return_value = True
+        b.playback.play.return_value.get.return_value = True
+
+        track1 = Track(uri='dummy:a', length=40000)
+        track2 = Track(uri='dummy:b', length=40000)
+
+        c = core.Core(mixer=None, backends=[b])
+        c.tracklist.add([track1, track2])
+
+        c.playback.play()
+        b.playback.change_track.assert_called_once_with(track1)
+        b.playback.change_track.reset_mock()
+
+        c.playback.pause()
+        c.playback.next()
+        b.playback.change_track.assert_called_once_with(track2)